+++ lib/src/environment.dart
@@ -35,6 +35,10 @@ class Environment {
     return this;
   }
 
+  String clientId() => 'wutsi-wallet';
+
+  int tenantId() => 1;
+
   String getGatewayUrl() => 'https://wutsi-gateway-$value.herokuapp.com';
 
   String getLoginUrl() =>  '${getGatewayUrl()}/login';
@@ -42,6 +46,8 @@ class Environment {
   String getOnboardUrl() => '${getLoginUrl()}/onboard';
 
   String getShellUrl() =>  '${getGatewayUrl()}/shell';
+
+  String getCashUrl() =>  '${getGatewayUrl()}/cash';
 
   String getStoreUrl() => '${getGatewayUrl()}/store';